chore: Remove unnecessary casts (#445)

+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/runtime/builder/MarkerUtil.java
@@ -128,7 +128,7 @@ public final class MarkerUtil {
     public static Long createdOn(IMarker marker, long onErrorValue) {
 
         try {
-            return (Long) marker.getCreationTime();
+            return marker.getCreationTime();
         } catch (CoreException ce) {
             return onErrorValue;
         }
